requirements.txt
# Main dependencies for the script
requests
apscheduler

# Dependencies for Telegram bot
python-telegram-bot